Add working drag and drop
Fixed an issue where I was setting vector of an object equal to the
vector of an event. This is a problem because Haxe passes by
reference, not value, and therefore the object would continue to
follow the mouse after being "turned off." Changed to lerping,
which fixed the issue.